+++ naccsweb/core/models.py
@@ -3,5 +3,5 @@ from django.db import models
 
 # Create your models here.
 
-class hubStats(models.Model):
+class HubStats(models.Model):
     matches = models.IntegerField()
